Fixed bitmap memory leak
Fixed an issue with the old bitmap function, where an array of size(width*height) would be initialized in SRAM. This caused images over 32x32 to break the function. The workaround was to eliminate all sort of temporary storage, and using some helper functions I fetch specific nibbles from PROGMEM when needed. 

Should also clean up the display if there is screen warping or flickering based on the redraw happening less often now.